move code back to indexjs, add comments
